fix: captcha token validation
